+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project starting with the 0.6.0 release will be docu
 ### Added
 - New field `frozen` appears in `sites list` when a site belonging to your user has been frozen. (#1015)
 - New command `site lookup` to look up sites by name. (#1027)
+- `site clear-cache` has new flags `--clear-framework-cache` and `--clear-varnish-cache`. (#1083)
 
 ### Changed
 - Removed the port number from the create-a-machine token URL seen when using `auth login` except for when the host is localhost. (#1034)